requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
itsdangerous==2.1.2